refactor(kbasehelper): rename and scope within selenium utilities
